Fri, Mar 4, 2022, 7:22 PM +00:00

+++ credit/credit.c
@@ -23,9 +23,9 @@ int main(void)
         printf ("INVALID\n");
     }
     int length = get_length(cc_number);
-    int power_1 = length - 1;
-    int first_digit = pow(cc_number, power_1);
-    printf("%i\n", first_digit);
+    double power_1 = length - 1;
+    double first_digit = pow(cc_number, power_1);
+    printf("%d\n", first_digit);
 
 
     //if (length)
